Buffer points before inserting them to InfluxDB
InfluxDB has WAL sync period set to 1s which means that serial insert throughput is only 1 insert/s and we can’t keep up with the MQTT topic if we insert each point separately. Thus buffer points in memory for 3 seconds or 1000 points and insert them as a batch to improve throughput even if a single insert request takes 1 second.